requirements.txt
streamlit
pyyaml
requests
spotipy>=2.23.0